model/recommender/pyproject.toml
[tool.poetry]
name = "artjuna-recommender"
version = "0.0.1"
description = ""
authors = ["Aditya Wardianto <83759122+ditw11mhs@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "3.9.7"
tensorflow = "^2.8.0"
numpy = "^1.22.3"
pandas = "^1.4.2"
python-dotenv = "^0.20.0"
Faker = "^13.12.0"
tqdm = "^4.64.0"
aiohttp = "^3.8.1"
requests-toolbelt = "^0.9.1"
fastparquet = "^0.8.1"

[tool.poetry.dev-dependencies]
mysql-connector-python = "^8.0.29"
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
